fix type bug

+++ train.py
@@ -85,7 +85,8 @@ def parse_record(raw_record, is_training, no_lmk):
     bbox_xmin = tf.maximum(tf.constant(0, dtype=tf.int64), bbox['xmin'])
     bbox_ymax = tf.minimum(tf.constant(_IMAGE_SIZE, dtype=tf.int64), bbox['ymax'])
     bbox_xmax = tf.minimum(tf.constant(_IMAGE_SIZE, dtype=tf.int64), bbox['xmax'])
-    crop_window = [bbox_ymin, bbox_xmin, bbox_ymax - bbox_ymin, bbox_xmax - bbox_xmin]
+    crop_window = tf.cast(
+        [bbox_ymin, bbox_xmin, bbox_ymax - bbox_ymin, bbox_xmax - bbox_xmin], dtype=tf.int32)
 
     cropped_image = tf.image.decode_and_crop_jpeg(image_buffer, crop_window, _NUM_CHANNELS)
